# src/main/java/com/bank/accounts/application/api/BalanceApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bank.accounts.application.api;

import com.bank.accounts.application.schemas.response.BadRequestResponse;
import com.bank.accounts.application.schemas.response.BalanceSchemaResponse;
import com.bank.accounts.application.schemas.response.InternalErrorResponse;
import com.bank.accounts.application.schemas.response.NotFoundResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Pattern;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Mono;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-20T01:45:15.001779300-05:00[America/Lima]")
@Validated
@Tag(name = "balance", description = "the balance API")
public interface BalanceApi {

    /**
     * GET /balance/payments/{accountId} : Obtener Total pagos del mes
     *
     * @param accountId (required)
     * @param month     (required)
     * @return (status code 200)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "getPaymentsBalance",
            summary = "Obtener Total pagos del mes",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "200", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BalanceSchemaResponse.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = InternalErrorResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/balance/payments/{accountId}",
            produces = {"application/json"}
    )
    Mono<BalanceSchemaResponse> getPaymentsBalance(
            @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") Long accountId,
            @NotNull @Pattern(regexp = "^\\d{4}-(0[1-9]|1[0-2])$") @Parameter(name = "month", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "month", required = true) String month,
            @Parameter(hidden = true) final ServerWebExchange exchange
    );


    /**
     * GET /balance/transactions/{accountId} : Obtener Total transacciones del mes
     *
     * @param accountId (required)
     * @param month     (required)
     * @return (status code 200)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "getTransactionsBalance",
            summary = "Obtener Total transacciones del mes",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "200", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BalanceSchemaResponse.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = InternalErrorResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/balance/transactions/{accountId}",
            produces = {"application/json"}
    )
    Mono<BalanceSchemaResponse> getTransactionsBalance(
            @Parameter(name = "accountId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accountId") Long accountId,
            @NotNull @Pattern(regexp = "^\\d{4}-(0[1-9]|1[0-2])$") @Parameter(name = "month", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "month", required = true) String month,
            @Parameter(hidden = true) final ServerWebExchange exchange
    );

}