# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: BANK-API-V1
  description: ''
  version: 1.0.0
...
paths:
...
  /balance/transactions/{accountId}:
    get:
      summary: Obtener Total transacciones del mes
      deprecated: false
      description: ''
      operationId: getTransactionsBalance
      tags: []
      parameters:
        - name: accountId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
        - name: month
          in: query
          description: ''
          required: true
          schema:
            type: string
            pattern: ^\d{4}-(0[1-9]|1[0-2])$
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceSchemaResponse'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
  /balance/payments/{accountId}:
    get:
      summary: Obtener Total pagos del mes
      deprecated: false
      description: ''
      operationId: getPaymentsBalance
      tags: []
      parameters:
        - name: accountId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
        - name: month
          in: query
          description: ''
          required: true
          schema:
            type: string
            pattern: ^\d{4}-(0[1-9]|1[0-2])$
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceSchemaResponse'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
components:
  schemas:
...
    BadRequestResponse:
      type: object
      properties:
        description:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - description
        - errors
    InternalErrorResponse: &ref_0
      type: object
      properties:
        description:
          type: string
      required:
        - description
...
    BalanceSchemaResponse:
      type: object
      properties:
        month:
          type: string
          pattern: ^\d{4}-(0[1-9]|1[0-2])$
        amount:
          type: string
      required:
        - month
        - amount
  responses:
...
    NotFound:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundResponse'
    BadRequest:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestResponse'
    InternalError:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalErrorResponse'
  securitySchemes: {}
servers: []
security: []